image-generation-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.9
requests==2.31.0
pillow==10.1.0
boto3==1.28.64